moved interpreter outside of src
